+++ src/pages/Home.tsx
@@ -4,9 +4,12 @@ import PortfolioList from "../components/portfolio/PortfolioList";
 import SuccessfulPortfolio from "../components/successfulPortfolio/SuccessfulPortfolio";
 import ProfessionalList from "../components/professional/ProfessionalList";
 import RecomendCreator from "../components/recommendCreator/RecommendCreator";
-import { Link, useSearchParams } from "react-router-dom";
+import { useSearchParams } from "react-router-dom";
 import { useState } from "react";
 import FilterPopup from "../components/filterPopup/FilterPopup";
+import FilterButton from "../components/home/FilterButton";
+import FilterTag from "../components/home/FilterTag";
+import RegisterInfo from "../components/home/RegisterInfo";
 
 const Home = (): JSX.Element => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
@@ -69,83 +72,6 @@ const Home = (): JSX.Element => {
         <RegisterInfo />
       </div>
     </PageLayout>
-  );
-};
-
-interface filterButtonProps {
-  children: string;
-  handleOpen: () => void;
-}
-
-const FilterButton = ({
-  children,
-  handleOpen,
-}: filterButtonProps): JSX.Element => {
-  return (
-    <div
-      className="text-xs gap-[4px] text-white bg-primary-50 p-[11px] rounded-full flex cursor-pointer"
-      onClick={handleOpen}
-    >
-      <img src="./filterWhite.svg" alt="filter" />
-      <span>{children}</span>
-    </div>
-  );
-};
-
-interface filterTagProps {
-  children: string;
-}
-
-const FilterTag = ({ children }: filterTagProps): JSX.Element => {
-  return (
-    <div className="text-xs gap-[4px] text-primary-70 p-[10px] border border-primary-70 rounded-full flex cursor-pointer">
-      <img src="./checkPrimary.svg" alt="check" />
-      <span>{children}</span>
-    </div>
-  );
-};
-
-const RegisterInfo = (): JSX.Element => {
-  return (
-    <div className="grid gap-s-3">
-      <div className="text-center text-neutral-70 text-2xl">
-        <p>회원가입 또는 로그인으로 SPECFOLIO</p>
-        <p>멋진 활동들을 펼쳐보세요.</p>
-      </div>
-      <div className="flex justify-center items-center gap-s-2">
-        <RegisterInfoButton to="/login" isLogin={true}>
-          로그인
-        </RegisterInfoButton>
-        <span className="text-neutral-40 text-xl">또는</span>
-        <RegisterInfoButton to="signup">회원가입</RegisterInfoButton>
-      </div>
-    </div>
-  );
-};
-
-interface registerInfoButtonProps {
-  children: string;
-  isLogin?: boolean;
-  to: string;
-}
-
-const RegisterInfoButton = ({
-  children,
-  isLogin,
-  to,
-}: registerInfoButtonProps) => {
-  const loginStyle = "text-neutral-90 border border-neutral-10";
-  const signupStyle = "bg-neutral-90 text-white";
-
-  return (
-    <Link
-      to={to}
-      className={`px-s-3 py-[10px] text-2xl rounded-full ${
-        isLogin ? loginStyle : signupStyle
-      }`}
-    >
-      {children}
-    </Link>
   );
 };
