+++ src/types/homeTypes.ts
@@ -1,0 +1,14 @@
+export interface registerInfoButtonProps {
+  children: string;
+  isLogin?: boolean;
+  to: string;
+}
+
+export interface filterTagProps {
+  children: string;
+}
+
+export interface filterButtonProps {
+  children: string;
+  handleOpen: () => void;
+}
